make sure we can support privately hosted repos
