+++ src/main/scala/com/gilcloud/sbt/gitlab/GitlabCredentials.scala
@@ -1,0 +1,17 @@
+package com.gilcloud.sbt.gitlab
+
+import sbt.DirectCredentials
+
+case class GitlabCredentials(key: String, value: String)
+
+object GitlabCredentials {
+
+  /**
+    * Apply method for Private-Token or Job-Token supplied by sbt credentials blob
+    * @param directCredentials credentials class
+    * @return GitlabCredentials
+    */
+  def apply(directCredentials: DirectCredentials): GitlabCredentials =
+    GitlabCredentials(directCredentials.userName, directCredentials.passwd)
+
+}
